+++ actionpack/test/fixtures/bad_helpers/typo_helper.rb
@@ -1,0 +1,2 @@
+module TaipoHelper
+end